requirements.txt
asgiref==3.5.2
autopep8==1.7.0
Django==4.1.1
django-appconf==1.0.5
django-ckeditor==6.5.0
django-compressor==4.1
django-jazzmin==2.5.0
django-js-asset==2.0.0
djangorestframework==3.13.1
Pillow==9.2.0
pycodestyle==2.9.1
pytz==2022.2.1
rcssmin==1.1.0
rjsmin==1.2.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.2
